pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "symparse"
version = "0.2.0"
description = "Self-optimizing Unix pipeline tool with neurosymbolic parsing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai==1.61.0",
    "google-re2==1.0.0",
    "jsonschema==4.23.0",
    "litellm>=1.50.0",
    "portalocker>=2.8.2"
]
classifiers = [
    "Environment :: Console",
    "Topic :: Text Processing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3"
]

[project.scripts]
symparse = "symparse.cli:main"
symparse-demo = "symparse.demo:main"

[project.optional-dependencies]
embed = [
    "sentence-transformers==3.4.1",
    "torch==2.5.1"
]
demo = ["asciinema"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
